Snittlistan.Web/Infrastructure/Installers/AutoMapperInstaller.cs
using AutoMapper;
using Castle.MicroKernel.Registration;
using Castle.MicroKernel.SubSystems.Configuration;
using Castle.Windsor;

namespace Snittlistan.Web.Infrastructure.Installers
{
    public class AutoMapperInstaller : IWindsorInstaller
    {
        public void Install(IWindsorContainer container, IConfigurationStore store)
        {
            container.Register(
                Classes
                    .FromThisAssembly()
                    .BasedOn<Profile>()
                    .LifestyleSingleton()
                    .WithServiceFromInterface(typeof(Profile)));
        }
    }
}